+++ src/css/ajax-panel.css
@@ -1,0 +1,3 @@
+.ajax-panel {
+  display: flex;
+}
