--- src/css/clock-panel.css
@@ -1,6 +1,0 @@
-.clock-panel {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-}
