+++ src/module.html
@@ -1,6 +1,6 @@
-<div class="clock-panel" ng-style="{height: ctrl.height}">
-  <span ng-if="ctrl.panel.dateSettings.showDate && ctrl.panel.mode === 'time'"
-    <h3 style="font-size: {{ctrl.panel.dateSettings.fontSize}}; font-weight: {{ctrl.panel.dateSettings.fontWeight}};">{{ctrl.date}}</h3>
-  </span>
-  <h2 style="font-size: {{ctrl.panel.timeSettings.fontSize}}; font-weight: {{ctrl.panel.timeSettings.fontWeight}};">{{ctrl.time}}</h2>
+<div class="ajax-panel" ng-style="{height: ctrl.height}">
+  BEFORE
+  <p class="panel-text-content" ng-bind-html="ctrl.content" ng-show="ctrl.content">
+  </p>
+  AFTER
 </div>
